requirements/test.txt
django-coverage-plugin~=1.3.0
django-debug-toolbar~=1.5.0
model-mommy~=1.2.0
pylint~=1.6.0
pylint-django~=0.7.0
pylint-plugin-utils~=0.2.0
pytest~=2.9.0
pytest-cov~=2.3.0
pytest-django~=2.9.1
pytest-mock~=1.0
pytest-pep8~=1.0.0
pytest-pylint~=0.6.0
pytest-testmon~=0.8.0
pytest-watch~=4.1.0
